Move the background work finished condition variable outside of the mutex
I mistook `GUARDED_BY` in the LevelDB code to mean that the condvar
also needed to be inside the mutex. Really it just needs a mutex guard to consume
in the `wait` method. Which means it's still guarded by a mutex. The
thread had to have a lock to pass in the mutex guard.