Fix: create new nodes when updating an existing value
When updating the value for an existing key, we remove the old cache
entry and insert a new entry with the updated value. This is so that
clients with a cache handle to the old entry keep a consistent view
and new requests get handles on the updated value.

There was some initial confusion about LevelDB cache behavior for
existing keys. Not sure why I thought the value wasn't being updated...