Only insert and compare user keys in the filter
This is so that we can find the most recent record for a key instead of
needing an exact match on sequence number and operation. I had a
misunderstanding about LevelDB deletion behavior but also I'm not
sure how LevelDB adjusts the filter policy to not depend on the sequence
number and operation. A [StackOverflow question](https://stackoverflow.com/questions/60600731/how-does-leveldb-handle-sequence-number-in-bloom-filter)
says that RocksDB will use `bits_per_key` to only process the user key
but I can't find the equivalent in LevelDB.