Fix: pass in the base version to apply changes to prevent leaking versions
I tried to encapsulate things by having the base version held within
the `VersionBuilder` but this was leaking the version. I was relying on
the builder being dropped before any `release_version` calls but this
was a bad implicit attempt at only needing to drop the `Arc` instead of
fully cleaning up the version.

It looks messy to pass in the version set and reach into it to drop the
verison node when the builder is dropped. I could pass the base version
in as a reference but I don't want to mess with lifetimes right now.