Use a `SyncSender` so that `CompactionWorker` can be marked `Sync`
This is necessary so that we can put `raindb::DB` in an `Arc` and share
it across multiple threads.

We set a limited buffer size but we empty it completely into the
unbounded, internal task buffer so this the task queue is effectively
infinite--just as it was previously. We add some warnings about the
queue filling up. Theoretically, the buffer filling should not be a
problem because RainDB slows down and pauses writes when compactions
are necessary (see `raindb::DB::make_room_for_write`).